Ads added + fix distance chart + restyle

+++ app/src/main/java/com/mini/infotainment/activities/stats/tab/StatsWeek.kt
@@ -66,7 +66,7 @@ class StatsWeek(override val ctx: ActivityStats) : StatsTab() {
         val title = ctx.getString(R.string.dist_trav)
         val description =
             ctx.getString(R.string.total_km)
-                .replace("{tot}", totalDistTrav.toInt().toString())
+                .replace("{tot}", totalDistTrav.toString())
 
         super.addChart(
             StatsChart(
